Fix potential IOError with DummyMP
There was an issue with DummyMP when many processes were
created, causing IOError to trigger occasionally due to going
over the limit for open file descriptors. This attempts to
work around the issue by only creating Queues when the process
is started, limiting the number of Queues to the number of
CPUs (or CPU cores) on the system.